Add BedTabixAdapter
