+++ packages/bed/src/BedTabixAdapter/declare.d.ts
